+++ src/main/scala/ai/starlake/job/ingest/loaders/BigQueryNativeLoader.scala
@@ -60,7 +60,6 @@ class BigQueryNativeLoader(ingestionJob: IngestionJob, accessToken: Option[Strin
           accessToken = accessToken
         )
       if (twoSteps) {
-        val startTime = System.currentTimeMillis()
         val (loadResults, tempTableIds, tableInfos) =
           ParUtils
             .runInParallel(settings.appConfig.maxParTask, path.map(_.toString).zipWithIndex) {
@@ -136,8 +135,6 @@ class BigQueryNativeLoader(ingestionJob: IngestionJob, accessToken: Option[Strin
           }
         }
 
-        println("First step done in : " + ExtractUtils.toHumanElapsedTimeFrom(startTime))
-
         val output: Try[List[BqLoadInfo]] =
           applyBigQuerySecondStep(
             targetConfig,
@@ -158,7 +155,9 @@ class BigQueryNativeLoader(ingestionJob: IngestionJob, accessToken: Option[Strin
           archiveTableTask(database, schema, table, info).foreach(_.run())
         }
         Try(ParUtils.runInParallel(settings.appConfig.maxParTask, tempTableIds) { tableId =>
-          new BigQueryNativeJob(targetConfig, "").dropTable(tableId)
+          BigQueryJobBase.recoverBigqueryException {
+            new BigQueryNativeJob(targetConfig, "").dropTable(tableId)
+          }
         })
           .flatMap(_ => output)
           .recoverWith { case exception =>